Add the home page and three recipe pages
The website started off blank and needed to be built out.
To fix this, I added the home, greek pasta salad, lobster rolls, and
simple egg salad pages to the website.

These pages needed to have information added to each of them, so
images, descriptions, ingredients, and steps were all added to each
recipe page.

Links were added to each recipe from the home page as well as links
to the home page were added on both the top and bottom of each recipe
page.